requirements.txt
streamlit==1.32.0
google-genai==0.5.0
librosa==0.10.1
numpy==1.26.4
soundfile==0.12.1
